Refactor GUI and add field manager
